requirements.txt
absl-py==2.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
av==16.0.1
bcrypt==4.3.0
billiard==4.2.4
boxsdk==3.9.2
celery==5.6.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==46.0.3
ctranslate2==4.6.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
ecdsa==0.19.1
exceptiongroup==1.3.1
executing==2.2.0
fastapi==0.121.3
faster-whisper==1.2.1
filelock==3.18.0
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2025.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.23.1
gym-notices==0.0.8
h11==0.16.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.6
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.8.0
kiwisolver==1.4.8
kombu==5.6.1
libclang==18.1.1
lxml==6.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpi4py==4.0.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.0
onnxruntime==1.23.2
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
paddlepaddle==2.6.2
pandas==1.4.3
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pybullet==3.2.6
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyparsing==3.2.1
pyperclip==1.8.2
pyproject-toml==0.1.0
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2022.2.1
PyYAML==6.0.2
pyzmq==27.0.0
redis==7.1.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
rubicon-objc==0.4.7
scipy==1.12.0
selenium==4.18.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.142
ultralytics-thop==2.0.14
urllib3==2.2.1
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.3.2
